refactor: separate TodoDao to 4 kinds of Dao
