+++ data/src/main/java/com/paraooo/data/local/entity/TodoEntity.kt
@@ -13,41 +13,38 @@ data class TodoTemplate(
     val title: String,
     val description: String,
     val hour: Int?, // null이면 시간 미지정,
-    val minute: Int?,
+    val minute: Int?, // null이면 시간 미지정
     @ColumnInfo(index = true) val type: TodoType, // GENERAL, PERIOD, DAY_OF_WEEK,
-    val alarmType : AlarmType
+    val alarmType : AlarmType // OFF, NOTIFY, POPUP
 )
 
-// ✅ 2. 특정 날짜에 생성된 Todo 인스턴스 (체크 상태 포함)
 @Entity(
     tableName = "todo_instance",
     foreignKeys = [
         ForeignKey(entity = TodoTemplate::class, parentColumns = ["id"], childColumns = ["templateId"], onDelete = CASCADE)
     ],
-    indices = [Index(value = ["date"]), Index(value = ["templateId"])] // 날짜와 템플릿 ID 인덱싱
+    indices = [Index(value = ["date"]), Index(value = ["templateId"])] // date, templateId 인덱스 설정
 )
 data class TodoInstance(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val templateId: Long, // 원본 TodoTemplate ID
-    val date: Long, // Todo가 속한 특정 날짜
-    val progressAngle: Float = 0F, // 체크 각도
+    val templateId: Long, // 원본 TodoTemplate Id
+    val date: Long,
+    val progressAngle: Float = 0F,
 )
 
-//// ✅ 3. 특정 기간 동안 반복되는 Todo 정보
 @Entity(
     tableName = "todo_period",
     foreignKeys = [
         ForeignKey(entity = TodoTemplate::class, parentColumns = ["id"], childColumns = ["templateId"], onDelete = CASCADE)
     ],
-    indices = [Index(value = ["templateId"])]
+    indices = [Index(value = ["templateId"])] // templateId 인덱스 설정
 )
 data class TodoPeriod(
-    @PrimaryKey val templateId: Long, // TodoTemplate의 ID
+    @PrimaryKey val templateId: Long, // 원본 TodoTemplate Id
     val startDate: Long,
     val endDate: Long
 )
 
-// ✅ 4. 특정 요일에 반복되는 Todo 정보
 @Entity(
     tableName = "todo_day_of_week",
     foreignKeys = [
@@ -57,12 +54,11 @@ data class TodoPeriod(
 )
 data class TodoDayOfWeek(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val templateId: Long, // TodoTemplate의 ID
+    val templateId: Long, // 원본 TodoTemplate Id
     val dayOfWeeks: List<Int>,
     val dayOfWeek: Int // 1(월) ~ 7(일)
 )
 
-// ✅ 5. Todo 유형 정의
 enum class TodoType {
     GENERAL, PERIOD, DAY_OF_WEEK
 }
